add sound control hook depending on volume and on/off
